[BE] REFACTOR: 도메인 내의 update 함수가 의존하는 dto를 도메인 레이어로 위치 변경 #1788
